R#: Moved GetCoreFx20Preview1 overloads and SplitFeedIntoLines methods into Feed class
