+++ FeedChecker/Feed.cs
@@ -1,4 +1,8 @@
-﻿namespace FeedChecker
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace FeedChecker
 {
     public class Feed
     {
@@ -8,5 +12,30 @@
         }
 
         public string Result { get; private set; }
+
+        public static IEnumerable<string> GetCoreFx20Preview1(Feed feed)
+        {
+            var lines = SplitFeedIntoLines(feed);
+
+            return GetCoreFx20Preview1(lines);
+        }
+
+        private static IEnumerable<string> GetCoreFx20Preview1(string[] lines)
+        {
+            var onlyCoreFxPreview1ButWithPackgeInFront =
+                lines.Where(l => l.StartsWith("Package: dotnet-hostfxr-2.0.0-preview1"));
+
+            // imagine there is more code
+
+            return onlyCoreFxPreview1ButWithPackgeInFront.Select(l => l.Replace("Package: ", ""));
+        }
+
+        private static string[] SplitFeedIntoLines(Feed feed)
+        {
+            var lines = feed.Result.Split(
+                new[] {"\r\n", "\r", "\n"},
+                StringSplitOptions.None);
+            return lines;
+        }
     }
 }
